Merge branches 'TomFLaS' and 'master' of https://github.com/Msmith6/FlagRunners into TomFLaS
Now contains Settings menu made by collaborator in the navigation menu.

Changed the Navigation menu; reduced the menu items and changed the header

When the user presses the Back button on the map screen, it acts as exiting the application rather than logging them out.

Added a working statistics page for the user.  Shows the user's username, flags captured, flags deployed, distance traveled, flags captured the most from, and flags captured the most by.

Added incrementation of user's "Flag Deployed" stat when the user deploys a flag.

Added the ability to log out of current user and return to the login screen.